WIP -- Move RaftCoordinator to RaftServer, and RaftServer to RaftState
